// src/types/slides.ts
export const enum ShapePathFormulasKeys {
  ROUND_RECT = 'roundRect',
  ROUND_RECT_DIAGONAL = 'roundRectDiagonal',
  ROUND_RECT_SINGLE = 'roundRectSingle',
  ROUND_RECT_SAMESIDE = 'roundRectSameSide',
  CUT_RECT_DIAGONAL = 'cutRectDiagonal',
  CUT_RECT_SINGLE = 'cutRectSingle',
  CUT_RECT_SAMESIDE = 'cutRectSameSide',
  CUT_ROUND_RECT = 'cutRoundRect',
  MESSAGE = 'message',
  ROUND_MESSAGE = 'roundMessage',
  L = 'L',
  RING_RECT = 'ringRect',
  PLUS = 'plus',
  TRIANGLE = 'triangle',
  PARALLELOGRAM_LEFT = 'parallelogramLeft',
  PARALLELOGRAM_RIGHT = 'parallelogramRight',
  TRAPEZOID = 'trapezoid',
  BULLET = 'bullet',
  INDICATOR = 'indicator',
}
// ...
export interface Gradient {
  type: GradientType
  colors: GradientColor[]
  rotate: number
}

export type LineStyleType = 'solid' | 'dashed' | 'dotted'

/**
 * Element Shadow
 *
 * h: Horizontal offset
 *
 * v: Vertical offset
 *
 * blur: Blur amount
 *
 * color: Shadow color
 */
export interface PPTElementShadow {
  h: number
  v: number
  blur: number
  color: string
}

/**
 * Element Border
 *
 * style?: Border style (solid or dashed)
 *
 * width?: Border width
 *
 * color?: Border color
 */
export interface PPTElementOutline {
  style?: LineStyleType
  width?: number
  color?: string
}
// ...
interface PPTBaseElement {
  id: string
  left: number
  top: number
  lock?: boolean
  groupId?: string
  width: number
  height: number
  rotate: number
  link?: PPTElementLink
  name?: string
}

export type TextType = 'title' | 'subtitle' | 'content' | 'item' | 'itemTitle' | 'notes' | 'header' | 'footer' | 'partNumber' | 'itemNumber'
// ...
export interface PPTTextElement extends PPTBaseElement {
  type: 'text'
  content: string
  defaultFontName: string
  defaultColor: string
  outline?: PPTElementOutline
  fill?: string
  lineHeight?: number
  wordSpace?: number
  opacity?: number
  shadow?: PPTElementShadow
  paragraphSpace?: number
  vertical?: boolean
  textType?: TextType
}
// ...
export interface ImageElementFilters {
  'blur'?: string
  'brightness'?: string
  'contrast'?: string
  'grayscale'?: string
  'saturate'?: string
  'hue-rotate'?: string
  'sepia'?: string
  'invert'?: string
  'opacity'?: string
}

export type ImageClipDataRange = [[number, number], [number, number]]

/**
 * Image Crop
 *
 * range: Crop range, e.g., [[10, 10], [90, 90]] means cropping the range from 10%, 10% to 90%, 90% of the original image from the top-left corner
 *
 * shape: Crop shape, see configs/imageClip.ts CLIPPATHS
 */
export interface ImageElementClip {
  range: ImageClipDataRange
  shape: string
}

export type ImageType = 'pageFigure' | 'itemFigure' | 'background'

/**
 * Image Element
 *
 * type: Element type (image)
 *
 * fixedRatio: Fixed image aspect ratio
 *
 * src: Image URL
 *
 * outline?: Border
 *
 * filters?: Image filters
 *
 * clip?: Crop information
 *
 * flipH?: Horizontal flip
 *
 * flipV?: Vertical flip
 *
 * shadow?: Shadow
 *
 * radius?: Corner radius
 *
 * colorMask?: Color mask
 *
 * imageType?: Image type
 */
export interface PPTImageElement extends PPTBaseElement {
  type: 'image'
  fixedRatio: boolean
  src: string
  outline?: PPTElementOutline
  filters?: ImageElementFilters
  clip?: ImageElementClip
  flipH?: boolean
  flipV?: boolean
  shadow?: PPTElementShadow
  radius?: number
  colorMask?: string
  imageType?: ImageType
}

export type ShapeTextAlign = 'top' | 'middle' | 'bottom'

/**
 * Text in Shape
 *
 * content: Text content (HTML string)
 *
 * defaultFontName: Default font (overridden by HTML inline styles in content)
 *
 * defaultColor: Default color (overridden by HTML inline styles in content)
 *
 * align: Text alignment direction (vertical direction)
 *
 * type: Text type
 */
export interface ShapeText {
  content: string
  defaultFontName: string
  defaultColor: string
  align: ShapeTextAlign
  type?: TextType
}
// ...
export interface PPTShapeElement extends PPTBaseElement {
  type: 'shape'
  viewBox: [number, number]
  path: string
  fixedRatio: boolean
  fill: string
  gradient?: Gradient
  pattern?: string
  outline?: PPTElementOutline
  opacity?: number
  flipH?: boolean
  flipV?: boolean
  shadow?: PPTElementShadow
  special?: boolean
  text?: ShapeText
  pathFormula?: ShapePathFormulasKeys
  keypoints?: number[]
}

export type LinePoint = '' | 'arrow' | 'dot'
// ...
export interface PPTLineElement extends Omit<PPTBaseElement, 'height' | 'rotate'> {
  type: 'line'
  start: [number, number]
  end: [number, number]
  style: LineStyleType
  color: string
  points: [LinePoint, LinePoint]
  shadow?: PPTElementShadow
  broken?: [number, number]
  broken2?: [number, number]
  curve?: [number, number]
  cubic?: [[number, number], [number, number]]
}

export type ChartType = 'bar' | 'column' | 'line' | 'pie' | 'ring' | 'area' | 'radar' | 'scatter'

export interface ChartOptions {
  lineSmooth?: boolean
  stack?: boolean
}

export interface ChartData {
  labels: string[]
  legends: string[]
  series: number[][]
}

/**
 * Chart Element
 *
 * type: Element type (chart)
 *
 * fill?: Fill color
 *
 * chartType: Basic chart type (bar/line/pie); all chart types are derived from these three basic types
 *
 * data: Chart data
 *
 * options: Extended options
 *
 * outline?: Border
 *
 * themeColors: Theme colors
 *
 * textColor?: Text color
 */
export interface PPTChartElement extends PPTBaseElement {
  type: 'chart'
  fill?: string
  chartType: ChartType
  data: ChartData
  options?: ChartOptions
  outline?: PPTElementOutline
  themeColors: string[]
  textColor?: string
}
// ...
export interface TableCell {
  id: string
  colspan: number
  rowspan: number
  text: string
  style?: TableCellStyle
}

/**
 * Table Theme
 *
 * color: Theme color
 *
 * rowHeader: Header row
 *
 * rowFooter: Footer row
 *
 * colHeader: First column
 *
 * colFooter: Last column
 */
export interface TableTheme {
  color: string
  rowHeader: boolean
  rowFooter: boolean
  colHeader: boolean
  colFooter: boolean
}

/**
 * Table Element
 *
 * type: Element type (table)
 *
 * outline: Border
 *
 * theme?: Theme
 *
 * colWidths: Array of column widths, e.g., [30, 50, 20] means the widths of the three columns are 30%, 50%, 20%
 *
 * cellMinHeight: Minimum cell height
 *
 * data: Table data
 */
export interface PPTTableElement extends PPTBaseElement {
  type: 'table'
  outline: PPTElementOutline
  theme?: TableTheme
  colWidths: number[]
  cellMinHeight: number
  data: TableCell[][]
}

/**
 * LaTeX Element (Formula)
 *
 * type: Element type (latex)
 *
 * latex: latex code
 *
 * path: svg path
 *
 * color: Color
 *
 * strokeWidth: Path width
 *
 * viewBox: SVG's viewBox attribute
 *
 * fixedRatio: Fixed shape aspect ratio
 */
export interface PPTLatexElement extends PPTBaseElement {
  type: 'latex'
  latex: string
  path: string
  color: string
  strokeWidth: number
  viewBox: [number, number]
  fixedRatio: boolean
}

/**
 * Video Element
 *
 * type: Element type (video)
 *
 * src: Video URL
 *
 * autoplay: Autoplay
 *
 * poster: Preview cover
 *
 * ext: Video extension, used to confirm resource type when the resource link is missing the suffix
 */
export interface PPTVideoElement extends PPTBaseElement {
  type: 'video'
  src: string
  autoplay: boolean
  poster?: string
  ext?: string
}

/**
 * Audio Element
 *
 * type: Element type (audio)
 *
 * fixedRatio: Fixed icon aspect ratio
 *
 * color: Icon color
 *
 * loop: Loop playback
 *
 * autoplay: Autoplay
 *
 * src: Audio URL
 *
 * ext: Audio extension, used to confirm resource type when the resource link is missing the suffix
 */
export interface PPTAudioElement extends PPTBaseElement {
  type: 'audio'
  fixedRatio: boolean
  color: string
  loop: boolean
  autoplay: boolean
  src: string
  ext?: string
}

export type PPTElement = PPTTextElement | PPTImageElement | PPTShapeElement | PPTLineElement | PPTChartElement | PPTTableElement | PPTLatexElement | PPTVideoElement | PPTAudioElement